+++ examples/article_results/fig04_convergence.py
@@ -16,6 +16,7 @@ from constants import *
 # Generate data and measure dimensions for different embedding dimensions (D), sample_sizes (n), 100 realization each
 # The dimensions computed with periodic boundary condition in the unit n-cube
 gen_and_measure = False  # if True, then generates dimension measurements and save it into Figure03_data.pkl (It takes hours)
+overwrite = False  # overwrites the "Figure04_data.pkl" data file if True
 
 Ds = [2, 6, 11, 13, 19, 26]
 ns = [10, 100, 500, 1000, 2500, 10000]
@@ -33,9 +34,9 @@ if gen_and_measure:
                           for j in range(realiz_id)]
                          for n in tqdm(ns)])
         id_m_dict[D] = id_m
-
-    with open(fn_data, 'wb') as f:
-        pickle.dump(id_m_dict, f)
+    if overwrite:
+        with open(fn_data, 'wb') as f:
+            pickle.dump(id_m_dict, f)
 else:
     with open(fn_data, 'rb') as f:
         id_m_dict = pickle.load(f)
